Added the marconi-core-json-rpc package.
* Moved inside it reusable functions from `marconi-tutorial` which help
  the creation of a JSON-RPC HTTP server using `servant-server`.

* Updated `marconi-tutorial` to use this new library
