+++ cabal.project
@@ -15,9 +15,10 @@ repository cardano-haskell-packages
 index-state:
   , hackage.haskell.org 2023-07-10T00:00:00Z
   , cardano-haskell-packages 2023-07-10T00:00:00Z
-packages: marconi-chain-index
+packages: legacy/marconi-core-legacy
+          marconi-chain-index
           marconi-core
-          legacy/marconi-core-legacy
+          marconi-core-json-rpc
           marconi-sidechain
           marconi-tutorial
