+++ marconi-core-json-rpc/README.md
@@ -1,0 +1,1 @@
+# marconi-core-json-rpc
