+++ marconi-core-json-rpc/changelog.d/scriv.ini
@@ -1,0 +1,1 @@
+../../scriv.ini
