+++ marconi-core-json-rpc/marconi-core-json-rpc.cabal
@@ -1,0 +1,77 @@
+cabal-version: 3.0
+name:          marconi-core-json-rpc
+version:       1.2.0.0
+synopsis:
+  Utilities for wrapping the query interface of Marconi indexers in a JSON-RPC HTTP server.
+
+description:
+  Please see the README on GitHub at <https://github.com/input-output-hk/marconi-core-json-rpc#README.md>
+
+category:      Testing
+homepage:
+  https://github.com/input-output-hk/marconi-core-json-rpc/README.md
+
+bug-reports:   https://github.com/input-output-hk/marconi/issues
+author:        IOHK
+maintainer:    operations@iohk.io
+license:       Apache-2.0
+license-files:
+  LICENSE
+  NOTICE
+
+build-type:    Simple
+
+source-repository head
+  type:     git
+  location: https://github.com/input-output-hk/marconi
+
+common lang
+  default-language:   Haskell2010
+  default-extensions:
+    DeriveAnyClass
+    DeriveGeneric
+    DerivingStrategies
+    ExplicitForAll
+    FlexibleContexts
+    FlexibleInstances
+    GADTs
+    GeneralizedNewtypeDeriving
+    ImportQualifiedPost
+    InstanceSigs
+    MultiParamTypeClasses
+    NamedFieldPuns
+    OverloadedStrings
+    PatternSynonyms
+    ScopedTypeVariables
+    StandaloneDeriving
+    Strict
+    TemplateHaskell
+    TupleSections
+    TypeApplications
+    TypeFamilies
+
+  ghc-options:
+    -Wall -Wnoncanonical-monad-instances -Wunused-packages
+    -Wincomplete-uni-patterns -Wincomplete-record-updates
+    -Wredundant-constraints -Widentities -Wmissing-import-lists
+
+library
+  import:          lang
+  exposed-modules: Marconi.Core.JsonRpc
+  hs-source-dirs:  src
+
+  ------------------------
+  -- Local dependencies
+  ------------------------
+  build-depends:
+    , marconi-chain-index:json-rpc
+    , marconi-core
+
+  ------------------------
+  -- Other dependencies
+  ------------------------
+  build-depends:
+    , base            >=4.7 && <5
+    , lens
+    , mtl
+    , servant-server
