+++ marconi-tutorial/marconi-tutorial.cabal
@@ -50,6 +50,7 @@ library
   build-depends:
     , marconi-chain-index:{marconi-chain-index, json-rpc}  >=1.2.0
     , marconi-core                                         >=1.2.0
+    , marconi-core-json-rpc                                >=1.2.0
 
   --------------------------
   -- Other IOG dependencies
