+++ marconi-tutorial/src/Marconi/Tutorial/HttpServer.hs
@@ -1,18 +1,18 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 
 module Marconi.Tutorial.HttpServer where
 
-import Control.Concurrent (MVar, readMVar)
-import Control.Lens (to, view, (^.))
-import Control.Monad.Except (ExceptT, runExceptT)
+import Control.Lens (to, view)
 import Control.Monad.IO.Class (liftIO)
 import Control.Monad.Reader (ReaderT, ask)
 import Data.Data (Proxy (Proxy))
 import Marconi.ChainIndex.Experimental.Indexers.Worker qualified as Core
 import Marconi.Core qualified as Core
+import Marconi.Core.JsonRpc qualified as Core
 import Marconi.Tutorial.CLI (optionsHttpPort)
 import Marconi.Tutorial.Env (
   Env,
@@ -21,10 +21,10 @@ import Marconi.Tutorial.Env (
  )
 import Marconi.Tutorial.Indexers.AddressCount (AddressCountQuery)
 import Network.JsonRpc.Server.Types ()
-import Network.JsonRpc.Types (JsonRpc, JsonRpcErr, RawJsonRpc, mkJsonRpcInternalErr)
+import Network.JsonRpc.Types (JsonRpc, RawJsonRpc)
 import Network.Wai.Handler.Warp (defaultSettings, runSettings, setPort)
 import Servant ((:>))
-import Servant.Server (Application, Handler, Server, serve)
+import Servant.Server (Application, Server, serve)
 
 {----------------------
    API specification
@@ -60,39 +60,4 @@ httpApp env = serve (Proxy @API) (httpServer env)
 httpServer
   :: Env
   -> Server API
-httpServer = getAddressCountHttpHandler
-
-getAddressCountHttpHandler
-  :: Env
-  -> AddressCountQuery
-  -> Handler (Either (JsonRpcErr String) (Core.Result AddressCountQuery))
-getAddressCountHttpHandler env =
-  queryIndexerVarHandler (env ^. addressCountIndexerWorker . to Core.standardWorkerIndexerVar)
-
-queryIndexerVarHandler
-  :: ( Show (Core.Result query)
-     , Ord (Core.Point event)
-     , Core.Queryable (ExceptT (Core.QueryError query) Handler) event query indexer
-     , Core.IsSync (ExceptT (Core.QueryError query) Handler) event indexer
-     )
-  => MVar (indexer event)
-  -> query
-  -> Handler (Either (JsonRpcErr String) (Core.Result query))
-queryIndexerVarHandler indexerVar q = do
-  indexer <- liftIO $ readMVar indexerVar
-  toHttpHandler indexer q
-
-toHttpHandler
-  :: ( Show (Core.Result query)
-     , Ord (Core.Point event)
-     , Core.Queryable (ExceptT (Core.QueryError query) Handler) event query indexer
-     , Core.IsSync (ExceptT (Core.QueryError query) Handler) event indexer
-     )
-  => indexer event
-  -> query
-  -> Handler (Either (JsonRpcErr String) (Core.Result query))
-toHttpHandler indexer q = do
-  countE <- runExceptT $ Core.queryLatest q indexer
-  case countE of
-    Left err -> pure $ Left $ mkJsonRpcInternalErr (Just $ show err)
-    Right count -> pure $ Right count
+httpServer = Core.queryHttpHandler (addressCountIndexerWorker . to Core.standardWorkerIndexerVar)
